Add index and recipes html files. Add CSS file
Initital setup of project in which an html file called
index.html was created to hold the content of the homepage.
Other html files called alfredo.html, gb.html and brownie.html
were created to hold the content for the each recipe. A CSS
file called style.css was created in order to be able to
add design to the project later on.