Add logo, img, link, para, recipe card to homepage
Made a homepage for the website consisting of a
interactable logo which takes you back to the homepage.
Added individual recipe cards for each recipe showing
an image, header and short paragraph describing each
dish. The image and header of the cards are interactable
which allow the users to go to the respected pages of
each recipe. Linked all pages for the recipe to the
homepage in index.